requirements.txt
Pillow
google
google.generativeai
pigment
PyQt6
ctypes
qtawesome
openai